requirements.txt
blapi-port
httpx
websockets
brotli
qrcode
aiohttp
python-dotenv